comments

+++ src/controllers/commentsController.js
@@ -1,19 +1,29 @@
 const commentModel = require('../model/Comments.js');
-const jwt = require('jsonwebtoken')
+const jwt = require('jsonwebtoken');
+
 class commentsController{
     static async listComments(req,res){
         try {
-            const comments = await commentModel.find();
-            res.status(200).send(comments);
+            const comments = await commentModel.find().populate({path: 'userId', select: 'name _id email'})
+                res.status(200).send(comments);
         } catch (error) {
-            res.status(500).send("Não foi possível pegar esses comentários " + error)
+            res.status(500).send("Não foi possível pegar esses comentários " + error);
         } 
     }
     static async createComment(req, res){
-        const comment = new commentModel(req.body);
+        const token = req.body.token;
+        const comment = new commentModel({
+            content: req.body.content,
+            postId: req.body.postId,
+            userId: req.body.userId,
+        });
         try{
-            await comment.save()
-            res.status(201).send("Comentário feito com sucesso");
+            const authorization = jwt.verify(token, process.env.SECRET_TOKEN);
+            if(req.body.userId != authorization._id) return res.status(403).send('não foi possível criar esse comentário');
+            else{
+                await comment.save();
+                res.status(201).send("Comentário feito com sucesso");
+            }
         }
         catch(err){
             res.status(500).send(`Não foi possível fazer o comentário:  ${err}`);
@@ -22,11 +32,14 @@ class commentsController{
     static async deleteComment(req,res){
         const id = req.params.id;
         try {
+            const findUser = await commentModel.findOne({_id: id});
+            const authorization = jwt.verify(req.body.token, process.env.SECRET_TOKEN);
+            if(findUser.userId._id != authorization._id) return res.status(403).send('Não foi possível deletar esse comentário')
             await commentModel.findByIdAndDelete(id);
             res.status(200).send("comentário excluido com sucesso!");
 
         } catch (error) {
-            res.status(500).send(`Não foi possível excluir esse comentário${error}`)
+            res.status(500).send(`Não foi possível excluir esse comentário${error}`);
         }
     }
 }
